@@ -6,7 +6,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.92.0+10
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=3.0.6 <4.0.0'
 
 dependencies:
   flutter:
@@ -15,23 +15,21 @@ dependencies:
     sdk: flutter
   provider: ^6.0.5
   cupertino_icons: ^1.0.2
-  cloud_firestore: ^3.5.1
-  cloud_functions: ^3.3.9
-  paginate_firestore: ^1.0.2
-  firebase_core: ^1.24.0
-  firebase_core_platform_interface: ^4.5.1
-  firebase_auth: ^3.11.2
-  firebase_storage: ^10.2.5
-  firebase_analytics: ^9.3.8
-  firebase_messaging: ^13.1.0
+  cloud_firestore: ^4.8.3
+  cloud_functions: ^4.3.4
+  # paginate_firestore: ^1.0.3+1
+  firebase_core: ^2.15.0
+  firebase_auth: ^4.7.1
+  firebase_storage: ^11.2.5
+  firebase_analytics: ^10.4.4
+  firebase_messaging: ^14.6.5
   sign_button: ^2.0.0
   sign_in_with_apple: ^4.3.0
-  google_fonts: ^3.0.1
+  google_fonts: ^4.0.4
   google_sign_in: ^5.0.4
   google_mobile_ads: ^2.1.0
   flutter_easyloading: ^3.0.5
   flutter_launcher_icons: ^0.11.0
-  flutter_icons: ^1.0.0
   flutter_secure_storage: ^7.0.1
   flutter_slidable: ^0.6.0
   flutter_udid: ^2.0.0
@@ -44,7 +42,6 @@ dependencies:
   package_info_plus: ^3.0.2
   image_picker: ^0.8.6+1
   rxdart: ^0.27.3
-  intl: ^0.17.0
   state_notifier: ^0.7.1
   device_info: ^2.0.3
   version: ^3.0.2
@@ -55,10 +52,22 @@ dependencies:
   audioplayers: ^3.0.1
   video_player: ^2.6.1
   algolia: ^1.1.2
+  flutter_native_splash: ^2.2.16
+  flutter_feather_icons: ^2.0.0+1
 
+flutter_native_splash:
+  image: 'images/icon.jpg' 
+  color: '#FCFAF2'
+  fullscreen: true
+  android_12:
+    icon_background_color: '#FCFAF2'
+    image: 'images/icon.jpg'
+    icon_background_color_dark: '#FCFAF2'
+    image_dark: 'images/icon.jpg'
 
 dependency_overrides:
-  firebase_core_platform_interface: 4.5.1
+  firebase_core_platform_interface: ^4.8.0
+  intl: ^0.18.1
   
  
 
