@@ -8,18 +8,19 @@ import 'all_thread.dart';
 // ignore: must_be_immutable
 class ListPage extends StatefulWidget {
   ListPage(
-      {Key key,
-      this.thread,
-      this.uid,
-      this.sort,
-      this.adInterstitial,
-      this.threadList})
-      : super(key: key);
+      {super.key,
+      required this.thread,
+      required this.uid,
+      required this.sort,
+      required this.adInterstitial,
+      required this.threadList,
+      required this.resSort});
   AdInterstitial adInterstitial;
   Thread thread;
   String uid;
   String sort;
   List<Thread> threadList;
+  bool resSort;
 
   @override
   // ignore: library_private_types_in_public_api
@@ -34,7 +35,7 @@ class _ListPageState extends State<ListPage>
   @override
   Widget build(BuildContext context) {
     super.build(context);
-    // widget.threadList.removeRange(0, 1);
+
     return Scaffold(
         backgroundColor: const Color(0xffFCFAF2),
         body: Container(
@@ -46,23 +47,25 @@ class _ListPageState extends State<ListPage>
                     sort: widget.sort,
                     uid: widget.uid,
                     adInterstitial: widget.adInterstitial,
-                    threadList: widget.threadList);
-                break;
+                    threadList: widget.threadList,
+                    resSort: widget.resSort);
+
               case '今日のトピック':
                 return TodayThread(
                     thread: widget.thread,
                     sort: widget.sort,
                     uid: widget.uid,
                     adInterstitial: widget.adInterstitial,
-                    threadList: widget.threadList);
-                break;
+                    threadList: widget.threadList,
+                    resSort: widget.resSort);
+
               default:
                 return ThreadList(
                     thread: widget.thread,
                     sort: widget.sort,
                     uid: widget.uid,
-                    adInterstitial: widget.adInterstitial);
-                break;
+                    adInterstitial: widget.adInterstitial,
+                    resSort: widget.resSort);
             }
           })(),
         ));
